+++ columbo/_types.py
@@ -25,6 +25,7 @@ Answer = Union[bool, str]
 Answers = Mapping[str, Answer]
 MutableAnswers = MutableMapping[str, Answer]
 OptionList = List[str]
+Options = Union[List[str], Mapping[str, str]]
 V = TypeVar("V")
 StaticOrDynamicValue = Union[V, Callable[[Answers], V]]
 ShouldAsk = Callable[[Answers], bool]